+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - **_Breaking Change_**: apply `anyhow` for better error handling.
+- **_Breaking Change_**: remove the implementation of `Deref/DerefMut` for `LambdaRuntimeApiClient`, implement `AsRef/AsMut` instead.
 - Feat: apply `tracing` instead of just `println` on error.
 
 ## v0.2.1